fix: pipe flushing mechanism
The previous code was checking the pipe's capacity, not its length! That
meant that the check for flushing the pipe was superfluous and doing
nothing, because of course the data in the pipe fits in the pipe, by
definition. I've instead implemented generic logging and always copy the
pipe data from the kernel to userland, then flush in userland instead.