feat: massive refactor to simplify interface a lot
This is a messy commit but all the changes came about together and would
be another huge chore to untangle. The essense of this commit though is
quite simple. It aims to redo the interface for polling which greatly
benefits the users of the interface. It also adds some extra cleanup
which wasn't being done before after all services have died